+++ pythran/tests/cases/train_eq.py
@@ -23,7 +23,7 @@ def train_eq(E, TrSyms, os, mu, wx,  errfctprs, adapt):
             if err[i].real*err[i-1].real > 0 and err[i].imag*err[i-1].imag > 0:
                 mu = mu
             else:
-                mu = mu/(1+mu*(err[i].real*err[i].real + err[i].imag*err[i].imag))
+                mu = np.float32(mu/(1+mu*(err[i].real*err[i].real + err[i].imag*err[i].imag)))
     return err, wx, mu
 
 def train_eq2(E, TrSyms, os, mu, wx,  errfctprs, adapt):
@@ -40,7 +40,7 @@ def train_eq2(E, TrSyms, os, mu, wx,  errfctprs, adapt):
             if err[i].real*err[i-1].real > 0 and err[i].imag*err[i-1].imag > 0:
                 mu = mu
             else:
-                mu = mu/(1+mu*(err[i].real*err[i].real + err[i].imag*err[i].imag))
+                mu = np.float32(mu/(1+mu*(err[i].real*err[i].real + err[i].imag*err[i].imag)))
     return err, wx, mu
 
 def cma_error(Xest, R, symbs):
